+++ internal/pkg/mapper/mapper_suite_test.go
@@ -1,0 +1,13 @@
+package mapper_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestLog(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Map Suite")
+}
