+++ internal/pkg/mapper/mapper_test.go
@@ -1,0 +1,15 @@
+package mapper
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+var _ = Describe("Zap options setup", func() {
+	BeforeEach(func() {
+	})
+
+	It("should set a custom writer", func() {
+		Expect(0).To(Equal(0))
+	})
+})
